ForgotPasswordActivity.java
Adds a forgot password section to the app.  The layout is a copy of the login page.  Thus it is currently using the login in button which at some point needs to be changed as that doesn't make sense.  Also the ability of going back to the main page.  It does actually send an email to the email typed in.  Note*** depending on which web browser you open the link within the email will cause the link to time out (not work).