upload DriverMapActivity.java
Virtually identical duplicate of RiderMapActivity.java but for the driver to find the rider.  Updates database with driver's location.  Depending on how we retrieve the driver rider combination either through a temporary key or using the others user ID.  All we would have to simply do is put the geo location of the other on the map as a marker.