Update RiderActivity.java
Changed the query to also pull the user rating and associates it with the map markers. This is on the assumption that the button will work like this: the user chooses a location first, then from the (available drivers) map markers the rider chooses their driver based on their rating by clicking on the map marker and accepting, then the (location) keys will be exchanged (which will allow the driver to see the rider).