Update RiderMapActivity
-Minor edits for demo.
Because the code is so nested it's difficult to comment out code, so here's some information provided about the current Map.
Info: 1. When putting a time, put XX00AM or PM where XX = current time hour or 1 hour ahead, this forces isAvailable.  If you don't it will horribly crash the app as it's not generate a random key for advanced booking.  Even if I added a random key it's not getting a preference time for the Driver so it still won't work.  So only demonstrate isAvailable.
2. This code currently poisons the database, i.e. stores values and then doesn't properly delete them.  It happens because it's needed for other things that need to be implemented or because some things aren't fully implemented.
3. Code runs in the realm of rainbows and butterflies, i.e. error checking hasn't been fully implemented.
4. So far I've discovered no errors in trying to run isAvailable or just by looking at the map.  Ryan has mentioned a bug where it keeps printing a Toast message about the time formatting being incorrect.  I believe this happens due the callback but, it doesn't occur on my Phone so I'm not entirely sure unless I see it myself.